P4ADEV-1545 FdRIngestionActity naming refactoring
