+++ src/main/java/it/gov/pagopa/payhub/activities/activity/reportingflow/SendEmailIngestionFlowActivity.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payhub.activities.activity.fdr;
+package it.gov.pagopa.payhub.activities.activity.reportingflow;
 
 /**
  * Interface for SendEmailIngestionFlowActivity.
