+++ src/main/java/it/gov/pagopa/payhub/activities/activity/reportingflow/service/IngestionFlowRetrieverService.java
@@ -1,0 +1,27 @@
+package it.gov.pagopa.payhub.activities.activity.reportingflow.service;
+
+import it.gov.pagopa.payhub.activities.dao.IngestionFlowDao;
+import it.gov.pagopa.payhub.activities.dto.reportingflow.IngestionFlowDTO;
+import it.gov.pagopa.payhub.activities.exception.IngestionFlowNotFoundException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+@Slf4j
+public class IngestionFlowRetrieverService {
+
+	private final IngestionFlowDao ingestionFlowDao;
+
+	public IngestionFlowRetrieverService(IngestionFlowDao ingestionFlowDao) {
+		this.ingestionFlowDao = ingestionFlowDao;
+	}
+
+	public IngestionFlowDTO getByFlowId(Long ingestionFlowId) {
+		Optional<IngestionFlowDTO> ingestionFlow = ingestionFlowDao.getIngestionFlow(ingestionFlowId);
+
+		return ingestionFlow
+			.orElseThrow(() -> new IngestionFlowNotFoundException("Cannot found ingestionFlow having id: "+ ingestionFlowId));
+	}
+}
