--- src/main/java/it/gov/pagopa/payhub/activities/dao/FlowHandlerDao.java
@@ -1,13 +1,0 @@
-package it.gov.pagopa.payhub.activities.dao;
-
-import it.gov.pagopa.payhub.activities.dto.fdr.FlowHandlerDTO;
-
-import java.util.Optional;
-
-public interface FlowHandlerDao {
-
-	/**
-	 *  * It will return the requested FlowHandlerDTO entity from its id
-	 * */
-	Optional<FlowHandlerDTO> getFlowHandler(Long ingestionFlowId);
-}
