+++ src/main/java/it/gov/pagopa/payhub/activities/dao/IngestionFlowDao.java
@@ -1,0 +1,13 @@
+package it.gov.pagopa.payhub.activities.dao;
+
+import it.gov.pagopa.payhub.activities.dto.reportingflow.IngestionFlowDTO;
+
+import java.util.Optional;
+
+public interface IngestionFlowDao {
+
+	/**
+	 *  * It will return the requested IngestionFlowDTO entity from its id
+	 * */
+	Optional<IngestionFlowDTO> getIngestionFlow(Long ingestionFlowId);
+}
