+++ src/main/java/it/gov/pagopa/payhub/activities/exception/IngestionFlowNotFoundException.java
@@ -1,0 +1,6 @@
+package it.gov.pagopa.payhub.activities.exception;
+
+public class IngestionFlowNotFoundException extends ActivitiesException {
+
+	public IngestionFlowNotFoundException(String message) { super(message); }
+}
